Significant changes! Back-compability broken
The creation of a model and training are now decoupled. This way, we can use the same CTC training script for many different model topologies. The create_model.lua should produce very similar results to the previous model.

Moved auxiliar lua files that are not main programs to src

BGRU/BiLSTM custom layers have been removed, they are part of cudnn.torch and the fixes that we made should be done there